Reorder methods/properties and extracted files to be compliant with the Stylecop Analyzer.
