+++ libraries/Microsoft.Bot.Builder.Dialogs/IDialogResume.cs
@@ -20,6 +20,7 @@ namespace Microsoft.Bot.Builder.Dialogs
         /// </summary>
         /// <param name="dc">The dialog context for the current turn of conversation.</param>
         /// <param name="result">(Optional) value returned from the dialog that was called. The type of the value returned is dependant on the dialog that was called.</param>
+        /// <returns>A <see cref="Task"/> representing the asynchronous operation.</returns>
         Task DialogResumeAsync(DialogContext dc, IDictionary<string, object> result);
     }
 }
