+++ libraries/Microsoft.Bot.Builder.Dialogs/Prompts/ChoicePrompt.cs
@@ -9,15 +9,6 @@ using static Microsoft.Bot.Builder.Dialogs.PromptValidatorEx;
 
 namespace Microsoft.Bot.Builder.Dialogs
 {
-    public class ChoicePromptOptions : PromptOptions
-    {
-        public List<Choice> Choices
-        {
-            get { return GetProperty<List<Choice>>(nameof(Choices)); }
-            set { this[nameof(Choices)] = value; }
-        }
-    }
-
     public class ChoicePrompt : Prompt<ChoiceResult>
     {
         private ChoicePromptInternal _prompt;
